+++ frontend/src/app/detectionEvents/page.tsx
@@ -1,0 +1,5 @@
+import { DetectionEventsHistory } from '@/components/detectionEvents/DetectionEventsHistory';
+
+export default function DetectionEventsPage() {
+  return <DetectionEventsHistory />;
+}
