--- frontend/src/app/events/page.tsx
@@ -1,5 +1,0 @@
-import { EventHistory } from '@/components/events/EventHistory';
-
-export default function EventsPage() {
-  return <EventHistory />;
-}
