+++ frontend/src/components/dashboard/DashboardMonitor.tsx
@@ -5,12 +5,12 @@ import { ChevronLeft, ChevronRight } from 'lucide-react';
 import { SystemMetrics } from './SystemMetrics';
 import { useAppSelector, useAppDispatch } from '@/store/hook';
 import { switchRole } from '@/store/slices/userSlice';
-import { CCTVGrid} from '@/components/monitoring/CCTVGrid';
-import { CameraDetailModal} from '@/components/monitoring/CameraDetailModal';
+import { CCTVGrid } from '@/components/monitoring/CCTVGrid';
+import { CameraDetailModal } from '@/components/monitoring/CameraDetailModal';
 import { FullscreenView } from '@/components/monitoring/FullscreenView';
-import { useGetCamerasQuery} from '@/services/apis/camera.api';
-import { useGetEventsQuery, useUpdateEventStatusMutation} from '@/services/apis/event.api';
-import { EventDetailModal } from '@/components/events/EventDetailModal';
+import { useGetCamerasQuery } from '@/services/apis/camera.api';
+import { useGetEventsQuery, useUpdateEventStatusMutation } from '@/services/apis/event.api';
+import { DetectionEventsDetailModal } from '@/components/detectionEvents/DetectionEventsDetailModal';
 import { useGetDashboardMetricsQuery } from '@/services/apis/metrics.api';
 import { RightPanel } from '@/components/monitoring/RightPanel';
 import { TopBar } from '@/layout/TopBar';
@@ -21,7 +21,7 @@ import { rolePermissions } from '@/types/anomaly';
 
 const KINDERGARTEN_ID = '1';
 
-// 하이브리드 모드용 가짜 데이터 (API 실패 시 사용)
+// 원본 그대로 유지된 하이브리드 모드용 가짜 데이터
 const allCameras: Camera[] = [
   { id: 'CAM-001', name: '정문 입구', location: '1층 현관', status: 'online', isRecording: true, category: 'entrance' },
   { id: 'CAM-002', name: '놀이터', location: '야외 놀이공간', status: 'online', isRecording: true, category: 'playground' },
@@ -46,8 +46,9 @@ const generateMockEvents = (): AnomalyEvent[] => {
     { id: 'EVT-001', timestamp: new Date(Date.now() - 2 * 60000), cameraId: 'CAM-001', cameraName: '정문 입구', type: 'Trespass', confidence: 95, location: '1층 현관', status: 'active', severity: 'high' },
     { id: 'EVT-002', timestamp: new Date(Date.now() - 15 * 60000), cameraId: 'CAM-009', cameraName: '운동장', type: 'Fight', confidence: 92, location: '야외 운동공간', status: 'reviewing', severity: 'high' },
     { id: 'EVT-003', timestamp: new Date(Date.now() - 45 * 60000), cameraId: 'CAM-002', cameraName: '놀이터', type: 'Wander', confidence: 78, location: '야외 놀이공간', status: 'resolved', severity: 'low' }
-  ] as AnomalyEvent[]).sort((a, b) => b.timestamp.getTime() - a.timestamp.getTime());
+  ] as AnomalyEvent[]).sort((a, b) => new Date(b.timestamp).getTime() - new Date(a.timestamp).getTime());
 };
+
 export function DashboardMonitor() {
   const dispatch = useAppDispatch();
   const { user } = useAppSelector((state) => state.user);
@@ -55,7 +56,6 @@ export function DashboardMonitor() {
   // RTK Query를 이용한 10초 주기 실시간 자동 갱신
   const { data: serverCameras, isError: isCameraError } = useGetCamerasQuery(KINDERGARTEN_ID, { skip: !user });
   const { data: serverEvents, isError: isEventError } = useGetEventsQuery({ kindergartenId: KINDERGARTEN_ID }, { skip: !user, pollingInterval: 10000 });
-  //const { data: serverEvents, isError: isEventError } = useGetEventsQuery(KINDERGARTEN_ID, { skip: !user, pollingInterval: 10000 });
   const { data: metrics, isError: isMetricsError } = useGetDashboardMetricsQuery(undefined, { skip: !user, pollingInterval: 10000 });
   const [updateEventStatus] = useUpdateEventStatusMutation();
 
@@ -97,24 +97,28 @@ export function DashboardMonitor() {
     }
   }, [serverEvents, isEventError]);
 
-  if (!user) return null;
+  if (!user) return <div className="h-screen flex items-center justify-center">Loading...</div>;
   const permissions = rolePermissions[user.role];
 
   // UI 핸들러들
   const handleEventStatusChange = async (status: AnomalyEvent['status']) => {
     if (selectedEvent) {
+      // 낙관적 업데이트: 사용자 경험을 위해 API 응답 대기 없이 UI를 먼저 즉시 반영
+      setLocalEvents(prev => prev.map(e => e.id === selectedEvent.id ? { ...e, status } : e));
+      setSelectedEvent(prev => prev ? { ...prev, status } : null);
+
       try {
-        if (!selectedEvent.id.includes('MOCK')) {
-          //await updateEventStatus({ eventId: selectedEvent.id, status }).unwrap();
+        if (!selectedEvent.id.includes('EVT')) {
           await updateEventStatus({ kindergartenId: KINDERGARTEN_ID, eventId: selectedEvent.id, status }).unwrap();
         }
       } catch (error) {
-        console.warn('상태 업데이트 실패. 로컬 화면만 반영됩니다.');
-      } finally {
-        setLocalEvents(prev => prev.map(e => e.id === selectedEvent.id ? { ...e, status } : e));
-        if (status === 'resolved') setSelectedEvent(null);
+        console.warn('API 상태 업데이트 실패. 로컬 화면만 반영됩니다.');
       }
     }
+  };
+
+  const handleAddNote = (note: string) => {
+    alert(`해당 이상상황에 메모가 저장되었습니다: ${note}`);
   };
 
   const cameraStats = {
@@ -130,54 +134,90 @@ export function DashboardMonitor() {
   const displayedCameras = filteredCameras.slice(startIndex, startIndex + camerasPerPage);
 
   return (
-    <>
-      <div className="h-screen flex flex-col bg-gray-50">
-        <TopBar currentRole={user.role} username={user.name} onRoleChange={(r) => dispatch(switchRole(r))} />
+      <>
+        <div className="h-screen flex flex-col bg-gray-50">
+          <TopBar currentRole={user.role} username={user.name} onRoleChange={(r) => dispatch(switchRole(r))} />
 
-        <div className="flex-1 flex overflow-hidden">
-          <Sidebar currentRole={user.role} userName={user.name} cameraStats={cameraStats} onCategoryFilter={setCategoryFilter} currentCategory={categoryFilter} />
+          <div className="flex-1 flex overflow-hidden">
+            <Sidebar currentRole={user.role} userName={user.name} cameraStats={cameraStats} onCategoryFilter={setCategoryFilter} currentCategory={categoryFilter} />
 
-          <main className="flex-1 p-6 overflow-auto">
-            <div className="mb-4 flex items-center justify-between">
-              <div>
-                <h2 className="text-lg font-bold text-gray-900">실시간 모니터링</h2>
-                <p className="text-sm text-gray-500">
-                  전체 {filteredCameras.length}개 중 {startIndex + 1}-{Math.min(startIndex + camerasPerPage, filteredCameras.length)}번째 • {layout} {isVideoPaused && ' • 일시정지됨'}
-                </p>
+            <main className="flex-1 p-6 overflow-auto">
+              <div className="mb-4 flex items-center justify-between">
+                <div>
+                  <h2 className="text-lg font-bold text-gray-900">실시간 모니터링</h2>
+                  <p className="text-sm text-gray-500">
+                    전체 {filteredCameras.length}개 중 {startIndex + 1}-{Math.min(startIndex + camerasPerPage, filteredCameras.length)}번째 • {layout} {isVideoPaused && ' • 일시정지됨'}
+                  </p>
+                </div>
+
+                {totalPages > 1 && (
+                    <div className="flex items-center gap-2">
+                      <Button variant="outline" size="sm" onClick={() => setCurrentPage(p => Math.max(1, p - 1))} disabled={currentPage === 1}>
+                        <ChevronLeft className="h-4 w-4" />
+                      </Button>
+                      <span className="text-sm font-medium text-gray-600 w-12 text-center">{currentPage} / {totalPages}</span>
+                      <Button variant="outline" size="sm" onClick={() => setCurrentPage(p => Math.min(totalPages, p + 1))} disabled={currentPage === totalPages}>
+                        <ChevronRight className="h-4 w-4" />
+                      </Button>
+                    </div>
+                )}
               </div>
 
-              {totalPages > 1 && (
-                <div className="flex items-center gap-2">
-                  <Button variant="outline" size="sm" onClick={() => setCurrentPage(p => Math.max(1, p - 1))} disabled={currentPage === 1}><ChevronLeft className="h-4 w-4" /></Button>
-                  <span className="text-sm font-medium text-gray-600 w-12 text-center">{currentPage} / {totalPages}</span>
-                  <Button variant="outline" size="sm" onClick={() => setCurrentPage(p => Math.min(totalPages, p + 1))} disabled={currentPage === totalPages}><ChevronRight className="h-4 w-4" /></Button>
-                </div>
-              )}
-            </div>
+              <div className="mb-6">
+                <CCTVGrid
+                    cameras={displayedCameras}
+                    onCameraSelect={(id) => setSelectedCamera(filteredCameras.find(c => c.id === id) || null)}
+                    onCameraFullscreen={(id) => setFullscreenCamera(filteredCameras.find(c => c.id === id) || null)}
+                    layout={layout}
+                />
+              </div>
 
-            <div className="mb-6">
-              <CCTVGrid cameras={displayedCameras} onCameraSelect={(id) => setSelectedCamera(filteredCameras.find(c => c.id === id) || null)} onCameraFullscreen={(id) => setFullscreenCamera(filteredCameras.find(c => c.id === id) || null)} layout={layout} />
-            </div>
+              {/*<div className="border-t border-gray-200 pt-8 mt-4">*/}
+              {/*  <SystemMetrics metrics={metrics} isError={isMetricsError} />*/}
+              {/*</div>*/}
 
-            {/* 💡 기존의 길고 지저분했던 MetricGauge 함수와 렌더링 코드를 단 한 줄로 대체했습니다. */}
-            <div className="border-t border-gray-200 pt-8 mt-4">
-              <SystemMetrics metrics={metrics} isError={isMetricsError} />
-            </div>
+            </main>
 
-          </main>
-
-          <RightPanel
-            events={localEvents} onEventClick={(id) => setSelectedEvent(localEvents.find(e => e.id === id) || null)} currentRole={user.role}
-            onLayoutChange={setLayout} currentLayout={layout} isRecording={isRecording} onRecordingToggle={() => setIsRecording(!isRecording)}
-            onQuickFullscreen={() => { if (filteredCameras.length > 0) setFullscreenCamera(filteredCameras[0]) }}
-            onQuickPause={() => setIsVideoPaused(!isVideoPaused)} onQuickDownload={() => alert('다운로드')} isVideoPaused={isVideoPaused}
-          />
+            <RightPanel
+                events={localEvents}
+                onEventClick={(id) => setSelectedEvent(localEvents.find(e => e.id === id) || null)}
+                currentRole={user.role}
+                onLayoutChange={setLayout}
+                currentLayout={layout}
+                isRecording={isRecording}
+                onRecordingToggle={() => setIsRecording(!isRecording)}
+                onQuickFullscreen={() => { if (filteredCameras.length > 0) setFullscreenCamera(filteredCameras[0]) }}
+                onQuickPause={() => setIsVideoPaused(!isVideoPaused)}
+                onQuickDownload={() => alert('다운로드')}
+                isVideoPaused={isVideoPaused}
+            />
+          </div>
         </div>
-      </div>
 
-      {selectedCamera && <CameraDetailModal camera={selectedCamera} onClose={() => setSelectedCamera(null)} onFullscreen={() => { setFullscreenCamera(selectedCamera); setSelectedCamera(null); }} />}
-      {selectedEvent && <EventDetailModal event={selectedEvent} onClose={() => setSelectedEvent(null)} onStatusChange={handleEventStatusChange} onAddNote={(note) => alert(`메모: ${note}`)} canResolve={permissions.canResolveAnomaly} />}
-      {fullscreenCamera && <FullscreenView camera={fullscreenCamera} onClose={() => setFullscreenCamera(null)} />}
-    </>
+        {selectedCamera && (
+            <CameraDetailModal
+                camera={selectedCamera}
+                onClose={() => setSelectedCamera(null)}
+                onFullscreen={() => { setFullscreenCamera(selectedCamera); setSelectedCamera(null); }}
+            />
+        )}
+
+        {selectedEvent && (
+            <DetectionEventsDetailModal
+                event={selectedEvent}
+                onClose={() => setSelectedEvent(null)}
+                onStatusChange={handleEventStatusChange}
+                onAddNote={handleAddNote}
+                canResolve={permissions.canResolveAnomaly}
+            />
+        )}
+
+        {fullscreenCamera && (
+            <FullscreenView
+                camera={fullscreenCamera}
+                onClose={() => setFullscreenCamera(null)}
+            />
+        )}
+      </>
   );
 }
